move bucket_value
