+++ corehq/ex-submodules/casexml/apps/phone/restore.py
@@ -25,7 +25,7 @@ from casexml.apps.phone.restore_caching import AsyncRestoreTaskIdCache, RestoreP
 from casexml.apps.phone.tasks import get_async_restore_payload, ASYNC_RESTORE_SENT
 from casexml.apps.phone.utils import get_cached_items_with_count
 from corehq.toggles import EXTENSION_CASES_SYNC_ENABLED, LIVEQUERY_SYNC
-from corehq.util.datadog.utils import bucket_value, maybe_add_domain_tag
+from corehq.util.datadog.utils import maybe_add_domain_tag
 from corehq.util.metrics import metrics_counter, metrics_histogram
 from corehq.util.timer import TimingContext
 from memoized import memoized
